Fix IP-port bugs. New ARM board. New GCC 7.3.0.
  1. build:   adapt OS for new GCC 7.3.0;
  2. kernel:  support IO-port mapping for x86_32/64;
  3. kernel:  support new ARM board Xilinx Zynq zc706;
  4. kernel:  begin SMP supporting;
  5. app:     improve UART userspace driver;
  6. app:     support memory allocation for mem_pool on app startup.

+++ krn/arch.h
@@ -8,11 +8,12 @@
 #define ARCH_H
 
 #include "sys_types.h"
+//#include "arch_data.h"
 
 void arch_init();
 void arch_set_timer_va(long va);
-void arch_set_ksp(uintptr_t ksp);
-uintptr_t arch_get_ksp();
+void arch_set_ksp(long ksp);
+long arch_get_ksp();
 void arch_user_invoke();
 void arch_switch_cpu(addr_t* cur_ksp, addr_t nxt_ksp, word_t nxt_utcb);
 void arch_store_floats(void* float_frame);
